requirements.txt
streamlit==1.19.0
Pillow==9.2.0
opencv-python==4.7.0.68
numpy==1.24.1
